feat(core): add and refactor configuration files
