+++ .changeset/large-ways-admire.md
@@ -1,0 +1,5 @@
+---
+"@yamada-ui/core": minor
+---
+
+Added `color-mix` util function.
