+++ scripts/generate-css/transform-props.ts
@@ -149,6 +149,34 @@ export const transformProps: TransformProps = {
     { properties: "maxBoxSize", transform: "px" },
   ],
   bgClip: ["backgroundClip"],
+  colorMix: [
+    "color",
+    "fill",
+    "floodColor",
+    "lightingColor",
+    "stroke",
+    "background",
+    "backgroundColor",
+    "caretColor",
+    "accentColor",
+    "outlineColor",
+    "textDecorationColor",
+    "textEmphasisColor",
+    "outlineColor",
+    "scrollbarColor",
+    "borderColor",
+    "borderTopColor",
+    "borderLeftColor",
+    "borderBottomColor",
+    "borderRightColor",
+    "borderInlineColor",
+    "borderInlineStartColor",
+    "borderInlineEndColor",
+    "borderBlockColor",
+    "borderBlockStartColor",
+    "borderBlockEndColor",
+    "columnRuleColor",
+  ],
   gradient: ["backgroundImage"],
   animation: ["animation"],
   transform: ["transform"],
